+++ android/src/main/kotlin/de/buddyactivities/fluttermesibo/DelegatingMessageListener.kt
@@ -13,12 +13,18 @@ import android.util.Log
 import com.mesibo.api.Mesibo
 
 /**
- * A {@link Mesibo.MessageListener} that delegates to the corresponding method of a given
- * {@link MesiboBinding.MesiboMessageListener} that manages platform channel layer.
+ * An implementation for {@link Mesibo.MessageListener} and {@link Mesibo.ConnectionListener} that
+ * delegates to the corresponding method of a given {@link MesiboBinding.MesiboMessageListener}
+ * which manages platform channel layer.
+ *
+ * NOTE: For some reason we can only add a single listener instance through 'Mesibo.addListener(listener)'.
+ * Otherwise one of the listeners won't be called. Therefore, we need to implement all listeners in the
+ * same class and delegate to the different target listeners that bridge to Flutter.
  */
 class DelegatingMessageListener(
-        private val targetListener: MesiboBinding.MesiboMessageListener,
-        private val modelMapper: BindingModelMapper) : Mesibo.MessageListener {
+        private val targetMessageListener: MesiboBinding.MesiboMessageListener,
+        private val targetConnectionListener: MesiboBinding.MesiboConnectionListener,
+        private val modelMapper: BindingModelMapper) : Mesibo.MessageListener, Mesibo.ConnectionListener {
 
     companion object {
         private const val TAG = "MesiboMsgListnr"
@@ -37,7 +43,7 @@ class DelegatingMessageListener(
         mappedMessage.params = modelMapper.toBindingMessageParams(params)
         mappedMessage.data = data
         // Pass message to Flutter
-        targetListener.onMessage(mappedMessage) {}
+        targetMessageListener.onMessage(mappedMessage) {}
 
         return true
     }
@@ -52,7 +58,7 @@ class DelegatingMessageListener(
 
         val mappedParams = modelMapper.toBindingMessageParams(params)
         // Pass message status to Flutter
-        targetListener.onMessageStatus(mappedParams) {}
+        targetMessageListener.onMessageStatus(mappedParams) {}
     }
 
     override fun Mesibo_onActivity(p0: Mesibo.MessageParams?, p1: Int) {
@@ -65,5 +71,36 @@ class DelegatingMessageListener(
 
     override fun Mesibo_onFile(p0: Mesibo.MessageParams?, p1: Mesibo.FileInfo?) {
         TODO("Not yet implemented")
+    }
+
+    override fun Mesibo_onConnectionStatus(statusCode: Int) {
+
+        val status = MesiboBinding.ConnectionStatus()
+        status.code = statusCode.toLong()
+
+        // Pass status to Flutter
+        targetConnectionListener.onConnectionStatus(status) {}
+
+        // log status change
+        val statusName = getConnectionStatusStr(statusCode)
+        Log.i(TAG, "Mesibo connection status: $statusName ($statusCode)")
+    }
+
+    private fun getConnectionStatusStr(statusCode: Int): String {
+        return when (statusCode) {
+            Mesibo.STATUS_ACTIVITY -> "activity"
+            Mesibo.STATUS_AUTHFAIL -> "auth fail"
+            Mesibo.STATUS_CONNECTFAILURE -> "connection failure"
+            Mesibo.STATUS_CONNECTING -> "connecting"
+            Mesibo.STATUS_MANDUPDATE -> "mandupdate"
+            Mesibo.STATUS_NONETWORK -> "no network"
+            Mesibo.STATUS_OFFLINE -> "offline"
+            Mesibo.STATUS_ONLINE -> "online"
+            Mesibo.STATUS_SHUTDOWN -> "shutdown"
+            Mesibo.STATUS_SIGNOUT -> "signout"
+            Mesibo.STATUS_STOPPED -> "stopped"
+            Mesibo.STATUS_UNKNOWN -> "unknown"
+            else -> "unexpected status"
+        }
     }
 }